requirements.txt
requests==2.32.5
beautifulsoup4==4.13.4
python-dotenv==1.2.1
